MS-917 task: separate logic in fromat-answers.js
Move logic to separate folder for re-use in  application
