+++ apps/nrm/behaviours/format-answers.js
@@ -1,40 +1,7 @@
 'use strict';
 
-/**
- * Capitalise the first letter or the first letter of each word (Title Case)
- *
- * @param {Object} text - text to be capitalised
- * @param {bool} [all=false] - whether the first letter of each word is capitalised
- *
- * @returns {string} - capitalised text
- */
-const capitaliseText = (text, all = false) => {
-  const capitalisedText = text.charAt(0).toUpperCase() + text.slice(1);
-
-  if (all) {
-    var splitString = text.toLowerCase().split(' ');
-    for (var i = 0; i < splitString.length; i++) {
-      splitString[i] = splitString[i].charAt(0).toUpperCase() + splitString[i].substring(1);
-    }
-    return splitString.join(' ');
-  }
-
-  return capitalisedText;
-
-};
-
-/**
- * Remove dashes from text
- *
- * @param {Object} text - text to be formatted
- *
- * @returns {string} - the text without dashes
- */
-const removeDashesFromText = (text) => {
-  const strippedDashes = text.replace('-', ' ');
-
-  return strippedDashes;
-};
+const dataFormatter = require('../util/data-formatter');
+const { removeDashesFromText, capitaliseText } = dataFormatter;
 
 /**
  * Format session answers for display purposes
@@ -69,8 +36,14 @@ const formatAnswers = (req) => {
   }
 
   if (req.sessionModel.get('pv-gender')) {
+    let formattedPvGenderValue = removeDashesFromText(capitaliseText(req.sessionModel.get('pv-gender')));
+
+    if (formattedPvGenderValue === 'Unknown') {
+      formattedPvGenderValue = 'They do not identify as male or female';
+    }
+
     data = Object.assign({}, data, {
-      formattedPvGender: removeDashesFromText(capitaliseText(req.sessionModel.get('pv-gender'))),
+      formattedPvGender: formattedPvGenderValue,
     });
   }
 
@@ -119,9 +92,7 @@ const formatAnswers = (req) => {
 module.exports = superclass => class extends superclass {
   locals(req, res) {
     const superlocals = super.locals(req, res);
-
     const data = Object.assign({}, formatAnswers(req));
-
     const locals = Object.assign({}, superlocals, data);
 
     return locals;
